Sync weekly plan weights with progressive targets
